renamed assignment_card
